Cleans up the archiving
